Code works with new VEX interface
The code is compatible with the new interface for VEX files. Raw
contents are still accessed from some functions that I expect to
refactor in the near future. The interface will be adjusted
if needed in the future.

Added unit tests for the method that loads scan data, with detailed
checks on the contents for two scans of a target and a calibrator.

Unit tests properly cover the current functionality of the interface