Make displacement models experiment-independent
The constructor of the Displacement class would take the experiment
object as input, and save it as an attribute. The attribute was not
accessed by any of the derived classes, nor by the base class itself,
so I removed it.

This is part of an effort to make the flow inside pride less obscure,
and stiff, by refactoring the main types to make them independent of
the existence of an Experiment object.